SOC to ISCO crosswalk

The 2010 Standard Occupational Classification (SOC) and the International Standard Classification of Occupations (ISCO-08) are compared. To make the crosswalk more straightforward and hence more useful, the notion of parsimony was applied. This means that while a task completed in the SOC may appear in numerous ISCOs (or vice versa), the match in some of these instances is just coincidental and adds unneeded complexity. This function allows mapping of data from the 4 SOC groups to the 4 ISCO levels.

Usage

soc_isco_crosswalk(
  data,
  soc_lvl,
  isco_lvl,
  brkd_cols = NULL,
  indicator = FALSE
)

Arguments

data,

data.table with mandatory columns job and value

soc_lvl,

character taking values from soc_1 to soc_4

isco_lvl,

numeric between 1 and 4

brkd_cols,

character vector with col names of stratification variables

indicator,

Boolean indicating if data describe an indicator. If TRUE the mean value is computed, otherwise the sum by each breakdown group.

Value

data.table with the estimated values for the requested ISCO occupational level.

References

Hardy W, Keister R, Lewandowski P (2018). “Educational upgrading, structural change and the task composition of jobs in Europe.” Economics of Transition, 26(2), 201--231.

Examples

library(iscoCrosswalks)
library(data.table)
#from soc_3 group to ISCO level 1 occupations
path <- system.file("extdata", "soc_3_brkdwn_example.csv",
                    package = "iscoCrosswalks")
dat <- fread(path)
soc_isco_crosswalk(dat,
                   soc_lvl = "soc_3",
                   isco_lvl = 1,
                   brkd_cols = "gender")
#>     isco08                          preferredLabel gender     value
#>  1:      5               Service and sales workers   Male 82.142857
#>  2:      5               Service and sales workers Female 52.457143
#>  3:      3 Technicians and associate professionals   Male 27.600000
#>  4:      3 Technicians and associate professionals Female 17.300000
#>  5:      2                           Professionals   Male  8.400000
#>  6:      1                                Managers   Male  6.857143
#>  7:      4                Clerical support workers   Male  6.000000
#>  8:      2                           Professionals Female  6.000000
#>  9:      1                                Managers Female  5.142857
#> 10:      4                Clerical support workers Female  3.100000